+++ python/aws_lambda_powertools/middleware_factory/__init__.py
@@ -1,4 +1,4 @@
 """ Utilities to enhance middlewares """
-from .middleware_factory import lambda_handler_decorator
+from .factory import lambda_handler_decorator
 
 __all__ = ["lambda_handler_decorator"]
