+++ python/tests/functional/test_utils.py
@@ -3,7 +3,7 @@ from typing import Callable
 
 import pytest
 
-from aws_lambda_powertools.utils import lambda_handler_decorator
+from aws_lambda_powertools.middleware_factory import lambda_handler_decorator
 
 
 @pytest.fixture
